refactor(ssm): rename packed sequence helper module
